Setup common ForkJoinPool to use classloader with access to module code
